Rename RotationDeadzone to EulerAngleDeadzone and tidy+ document file
